+++ pyronear_mlops/model/yolo/configs/yolov12/baseline.yaml
@@ -1,0 +1,4 @@
+---
+epochs: 50
+imgsz: 640
+model_type: yolo12n.pt
