+++ pyronear_mlops/model/yolo/configs/yolov12/best.yaml
@@ -1,0 +1,15 @@
+---
+model_type: yolo12s.pt
+epochs: 40
+patience: 15
+batch: 8
+imgsz: 1024
+lr0: 0.013281489698569585
+lrf: 0.0015264179671752333
+optimizer: auto
+mixup: 0.2
+close_mosaic: 5
+degrees: 7.777777777777779
+translate: 0.08888888888888889
+flipud: 0.0
+fliplr: 0.5
