+++ scripts/model/yolo/configs/baseline.yaml
@@ -1,0 +1,4 @@
+---
+epochs: 50
+imgsz: 640
+model_type: yolo12n.pt
