--- scripts/model/yolov8/configs/baseline.yaml
@@ -1,3 +1,0 @@
-epochs: 50
-imgsz: 640
-model_type: yolov8n.pt
