--- scripts/model/yolov8/configs/best.yaml
@@ -1,17 +1,0 @@
----
-# Model type: yolov8n, yolov8s, yolov8m, etc
-model_type: yolov8s.pt
-# Hyperparameters: the best combination was found by performing random hyperparameter search
-epochs: 200
-patience: 100
-batch: 16
-imgsz: 1024
-lr0: 0.013281489698569585
-lrf: 0.0015264179671752333
-optimizer: auto
-mixup: 0.2
-close_mosaic: 3
-degrees: 7.777777777777779
-translate: 0.08888888888888889
-flipud: 0.0
-fliplr: 0.5
